// src/app/models/employement.ts
export interface EmployemntData {
    agrictulture_ratio: number;
    employed_percent: number;
    employed_total: number;
    footnotes: string;
    labor_force: string;
    nonagriculture_ratio: number;
    not_in_labor: number;
    population: number;
    population_percent: number;
    unemployed: number;
    unemployed_percent: number;
    year: number;
}